cleaned up formatters and parsers
