+++ src/System.Text.Primitives/System/Text/Formatters/Formatter.cs
