+++ src/System.Text.Primitives/System/Text/Formatters/Formatter_floats.cs
