+++ src/System.Text.Primitives/System/Text/Formatters/Utf16Formatter_guid.cs
