+++ src/System.Text.Primitives/System/Text/Parsers/Utf6Parser_bool.cs
