+++ src/System.Text.Primitives/System/Text/Parsers/Utf6Parser_signed_hex.cs
