+++ src/System.Text.Primitives/System/Text/SymbolTable/ParsingTrie.cs
