+++ src/System.Text.Primitives/System/Text/SymbolTable/SymbolTable.cs
