+++ src/System.Text.Primitives/System/Text/SymbolTable/SymbolTable_symbol.cs
