+++ src/System.Text.Primitives/System/Text/SymbolTable/SymbolTable_utf16.cs
