+++ src/System.Text.Primitives/System/Text/SymbolTable/SymbolTable_utf8.cs
