+++ tests/System.Text.Primitives.Tests/Formatting/BooleanTests.cs
@@ -16,22 +16,6 @@ namespace System.Text.Primitives.Tests
         [InlineData('l', false, "false")]
         [InlineData(' ', true, "True")]
         [InlineData(' ', false, "False")]
-        public void BooleanUtf8(char format, bool value, string expected)
-        {
-            StandardFormat f = format == ' ' ? default : new StandardFormat(format);
-            byte[] buffer = new byte[256];
-            Assert.True(Utf8Formatter.TryFormat(value, buffer, out int bytesWritten, f));
-            var actual = Text.Encoding.UTF8.GetString(buffer, 0, bytesWritten);
-            Assert.Equal(expected, actual);
-        }
-
-        [Theory]
-        [InlineData('G', true, "True")]
-        [InlineData('G', false, "False")]
-        [InlineData('l', true, "true")]
-        [InlineData('l', false, "false")]
-        [InlineData(' ', true, "True")]
-        [InlineData(' ', false, "False")]
         public void BooleanUtf16(char format, bool value, string expected)
         {
             StandardFormat f = format == ' ' ? default : new StandardFormat(format);
